noticias

+++ app/Http/Controllers/NoticiasController.php
@@ -2,15 +2,17 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 
 use App\Noticia;
 use App\NoticiaHistorial;
 use App\NoticiaComentario;
 use App\User;
+use Illuminate\Http\Request;
+use Laracasts\Flash\Flash;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Redirect;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
-use Laracasts\Flash\Flash;
 
 class NoticiasController extends Controller
 {
@@ -64,26 +66,29 @@ class NoticiasController extends Controller
 
         $user_id = \Auth::user()->id;
 
+        $noticia->load('user', 'noticiacomentarios', 'noticiahistorial');
+        $noticia->noticiahistorial->load('user');
+
 /*
         $noticiahistorialr = NoticiaHistorial::where('user_id',$user_id);
 
         $noticiahistorialr = DB::table('noticiahistorials')->where('user_id',$user_id)->first();
-*/
-        $noticiahistorialr = \DB::select('SELECT nh.id, nh.user_id
+        $noticiahistorialr = \DB::select('SELECT nh.id as historial, nh.user_id as id
                                             FROM noticiahistorials nh
                                             INNER JOIN noticias ns on ns.id = nh.noticia_id
                                             inner join users u on u.id = nh.user_id
-                                            WHERE nh.noticia_id = "{$id}" and u.id = "{$user_id}";');
+                                            WHERE nh.noticia_id = "{$id}" and u.id = "{$user_id}"');
 
 
-        foreach ($noticiahistorialr as $historial){
-/*
-            $historial->delete();
-
+        dd($noticiahistorialr);
 */
-            $hid = $hostial->id;
-            $noticiahistorialv = NoticiaHistorial::find($hid);
-            $noticiahistorialv->delete();
+
+        foreach ($noticia->noticiahistorial as $historial){
+            if ( $user_id == $historial->user_id){
+                $hid = $historial->id;
+                $noticiahistorialv = NoticiaHistorial::find($hid);
+                $noticiahistorialv->delete();
+            }
         }
 
         $noticiahistorial = new NoticiaHistorial();
@@ -92,10 +97,14 @@ class NoticiasController extends Controller
         $noticiahistorial -> save();
 
 
+
         $noticia->load('user', 'noticiacomentarios', 'noticiahistorial');
 
+        $users = User::orderBy('id', 'DESC')->paginate(20);
+
         return view('front.noticias.show')
-            ->with('noticia', $noticia);
+            ->with('noticia', $noticia)
+            ->with('users', $users);
     }
 
     /**
